# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/events":{"post":{"operationId":"EventController_createEvent","parameters":[],"responses":{"201":{"description":""}},"tags":["Event"]},"get":{"operationId":"EventController_getAllEvents","parameters":[],"responses":{"200":{"description":""}},"tags":["Event"]}},"/events/{id}":{"get":{"operationId":"EventController_getEventById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Event"]},"patch":{"operationId":"EventController_updateEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Event"]},"delete":{"operationId":"EventController_deleteEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Event"]}},"/events/{eventId}/performances":{"post":{"operationId":"PerformanceController_create","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePerformanceDto"}}}},"responses":{"201":{"description":"공연이 성공적으로 생성되었습니다."}},"security":[{"bearer":[]}],"summary":"공연 생성","tags":["공연"]}},"/performances":{"get":{"operationId":"PerformanceController_findAll","parameters":[],"responses":{"200":{"description":"모든 공연 목록을 반환합니다."}},"summary":"모든 공연 조회","tags":["공연"]}},"/performances/{id}":{"get":{"operationId":"PerformanceController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"공연 정보를 반환합니다."},"404":{"description":"공연을 찾을 수 없습니다."}},"summary":"공연 상세 조회","tags":["공연"]},"patch":{"operationId":"PerformanceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePerformanceDto"}}}},"responses":{"200":{"description":"공연이 성공적으로 수정되었습니다."}},"security":[{"bearer":[]}],"summary":"공연 수정","tags":["공연"]},"delete":{"operationId":"PerformanceController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"공연이 성공적으로 삭제되었습니다."}},"security":[{"bearer":[]}],"summary":"공연 삭제","tags":["공연"]}}},"info":{"title":"Fast Pass API","description":"The Fast Pass API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}},"CreatePerformanceDto":{"type":"object","properties":{"startAt":{"type":"string","example":"2025-01-01T20:00:00Z","description":"Performance start time"},"totalSeats":{"type":"number","example":150,"description":"Total number of seats"},"availableSeats":{"type":"number","example":150,"description":"Available seats"}},"required":["startAt","totalSeats"]},"UpdatePerformanceDto":{"type":"object","properties":{"startAt":{"type":"string","example":"2025-01-01T20:00:00Z","description":"Performance start time"},"totalSeats":{"type":"number","example":150,"description":"Total number of seats"},"availableSeats":{"type":"number","example":150,"description":"Available seats"}}}}}}